chore(UIForm): align folder architecture
