+++ packages/forms/src/UIForm/fields/CheckBox/index.js
@@ -1,0 +1,4 @@
+import CheckBox from './CheckBox.component';
+import CheckBoxes from './CheckBoxes.component';
+
+export { CheckBox, CheckBoxes };
