+++ packages/forms/src/UIForm/fields/Radios/index.js
@@ -1,0 +1,3 @@
+import Radios from './Radios.component';
+
+export default Radios;
