+++ packages/forms/src/UIForm/fields/Select/index.js
@@ -1,0 +1,3 @@
+import Select from './Select.component';
+
+export default Select;
