+++ packages/forms/src/UIForm/fields/Text/index.js
@@ -1,0 +1,3 @@
+import Text from './Text.component';
+
+export default Text;
