+++ packages/forms/src/UIForm/fields/TextArea/TextArea.component.test.js
@@ -1,21 +1,22 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import Text from './Text';
+import TextArea from './TextArea.component';
 
-describe('Text field', () => {
+describe('TextArea field', () => {
 	const schema = {
 		autoFocus: true,
 		description: 'my text input hint',
+		key: ['user', 'comment'],
 		placeholder: 'Type something here',
 		title: 'My input title',
 		type: 'text',
 	};
 
-	it('should render input', () => {
+	it('should render textarea', () => {
 		// when
 		const wrapper = shallow(
-			<Text
+			<TextArea
 				id={'myForm'}
 				isValid
 				errorMessage={'My error message'}
@@ -29,7 +30,7 @@ describe('Text field', () => {
 		expect(wrapper.node).toMatchSnapshot();
 	});
 
-	it('should render disabled input', () => {
+	it('should render disabled textarea', () => {
 		// given
 		const disabledSchema = {
 			...schema,
@@ -38,7 +39,7 @@ describe('Text field', () => {
 
 		// when
 		const wrapper = shallow(
-			<Text
+			<TextArea
 				id={'myForm'}
 				isValid
 				errorMessage={'My error message'}
@@ -52,7 +53,7 @@ describe('Text field', () => {
 		expect(wrapper.node).toMatchSnapshot();
 	});
 
-	it('should render readonly input', () => {
+	it('should render readonly textarea', () => {
 		// given
 		const readOnlySchema = {
 			...schema,
@@ -61,7 +62,7 @@ describe('Text field', () => {
 
 		// when
 		const wrapper = shallow(
-			<Text
+			<TextArea
 				id={'myForm'}
 				isValid
 				errorMessage={'My error message'}
@@ -75,11 +76,34 @@ describe('Text field', () => {
 		expect(wrapper.node).toMatchSnapshot();
 	});
 
+	it('should render provided rows', () => {
+		// given
+		const schemaWithRows = {
+			...schema,
+			rows: 10,
+		};
+
+		// when
+		const wrapper = shallow(
+			<TextArea
+				id={'myForm'}
+				isValid
+				errorMessage={'My error message'}
+				onChange={jest.fn()}
+				schema={schemaWithRows}
+				value={'toto'}
+			/>
+		);
+
+		// then
+		expect(wrapper.node).toMatchSnapshot();
+	});
+
 	it('should trigger onChange', () => {
 		// given
 		const onChange = jest.fn();
 		const wrapper = shallow(
-			<Text
+			<TextArea
 				id={'myForm'}
 				isValid
 				errorMessage={'My error message'}
@@ -88,38 +112,13 @@ describe('Text field', () => {
 				value={'toto'}
 			/>
 		);
-		const event = { target: { value: 'totoa' } };
+		const value = 'totoa';
+		const event = { target: { value } };
 
 		// when
-		wrapper.find('input').simulate('change', event);
+		wrapper.find('textarea').simulate('change', event);
 
 		// then
-		expect(onChange).toBeCalledWith(event, schema, 'totoa');
-	});
-
-	it('should trigger onChange with number value', () => {
-		// given
-		const numberSchema = {
-			...schema,
-			type: 'number',
-		};
-		const onChange = jest.fn();
-		const wrapper = shallow(
-			<Text
-				id={'myForm'}
-				isValid
-				errorMessage={'My error message'}
-				onChange={onChange}
-				schema={numberSchema}
-				value={2}
-			/>
-		);
-		const event = { target: { value: '25' } };
-
-		// when
-		wrapper.find('input').simulate('change', event);
-
-		// then
-		expect(onChange).toBeCalledWith(event, numberSchema, 25);
+		expect(onChange).toBeCalledWith(event, schema, value);
 	});
 });
