+++ packages/forms/src/UIForm/fields/TextArea/index.js
@@ -1,0 +1,3 @@
+import TextArea from './TextArea.component';
+
+export default TextArea;
