+++ packages/forms/src/UIForm/fieldsets/Fieldset/index.js
@@ -1,0 +1,3 @@
+import Fieldset from './Fieldset.component';
+
+export default Fieldset;
