+++ packages/forms/src/UIForm/fieldsets/Tabs/index.js
@@ -1,0 +1,3 @@
+import Tabs from './Tabs.component';
+
+export default Tabs;
