+++ packages/forms/src/UIForm/utils/widgets.js
@@ -1,10 +1,8 @@
 import Fieldset from '../fieldsets/Fieldset';
 import Tabs from '../fieldsets/Tabs';
 
-import Button from '../fields/Button/Button.component';
-import Buttons from '../fields/Button/Buttons.component';
-import CheckBox from '../fields/CheckBox/CheckBox.component';
-import CheckBoxes from '../fields/CheckBox/CheckBoxes.component';
+import { Button, Buttons } from '../fields/Button';
+import { CheckBox, CheckBoxes } from '../fields/CheckBox';
 import Radios from '../fields/Radios';
 import Select from '../fields/Select';
 import Text from '../fields/Text';
